🔥 Remove everything related to public plugin
This is just now an interface to handle the two requests (for login) required by Geovelo. The credentials are stored in the TRMNL plugin configuration, it’s much simpler.
I might have to add some calculations on data before sending it back to TRMNL, but no more storing things and making layouts outside of TRMNL’s editor.